// src/app/Models/Candidate.ts
export class candidate {
  buId: string;
  buName: string;
  candidateType: string;
  contactNumber: string;
  createdBy: string;
  createdByName: string;
  createdDateTime: string;
  currentCompany: string;
  customerName: string;
  email: string;
  employmentType: string;
  experience: string;
  firstName: string;
  fromDateTime: string;
  fullName: string;
  id: string;
  image: string;
  imageType: string;
  isBench: number;
  jrId: string;
  jrNumber: string;
  lastName: string;
  location: string;
  message: string;
  noOfOpenings: string;
  placementFor: string;
  recStatus: string;
  recStatusDisplay: string;
  remarks: string;
  resourceId: string;
  resumeName: string;
  resumeType: string;
  roleName: string;
  sex: string;
  skillSet: string;
  status: string;
  toDateTime: string;
  updatedBy: string;
  updatedByName: string;
  updatedDateTime: string;
  vendorId: string;
  vendorName: string;
  refId: string;
  keyword: string;
}